backend/requirements.txt
fastapi
uvicorn
huggingface_hub
pydantic
python-dotenv
openai-whisper
torch
